+++ src/commands/index.ts
@@ -1,59 +1,60 @@
 // Tenant
 export * from "./tenant/tenant.fetch.ts"
 export * from "./tenant/tenant.list.ts"
+export * from "./tenant/tenant.translate-name-to-id.ts"
 
 // Api Key
-export * from "./api-key/api-key.list.ts"
 export * from "./api-key/api-key.create.ts"
 export * from "./api-key/api-key.delete.ts"
+export * from "./api-key/api-key.list.ts"
 
 // Secret
-export * from "./secret/secret.list.ts"
 export * from "./secret/secret.create.ts"
 export * from "./secret/secret.delete.ts"
+export * from "./secret/secret.list.ts"
 
 // Variable
-export * from "./variable/variable.list.ts"
 export * from "./variable/variable.create.ts"
 export * from "./variable/variable.delete.ts"
+export * from "./variable/variable.list.ts"
 
 // Data Core
+export * from "./data-core/data-core.create.ts"
+export * from "./data-core/data-core.delete-request.ts"
 export * from "./data-core/data-core.exists.ts"
 export * from "./data-core/data-core.fetch.ts"
 export * from "./data-core/data-core.list.ts"
-export * from "./data-core/data-core.create.ts"
-export * from "./data-core/data-core.update.ts"
-export * from "./data-core/data-core.delete-request.ts"
 export * from "./data-core/data-core.request-delete.ts"
+export * from "./data-core/data-core.update.ts"
 
 // Flow Types
+export * from "./flow-type/flow-type.create.ts"
+export * from "./flow-type/flow-type.delete-request.ts"
 export * from "./flow-type/flow-type.exists.ts"
 export * from "./flow-type/flow-type.fetch.ts"
 export * from "./flow-type/flow-type.list.ts"
-export * from "./flow-type/flow-type.create.ts"
-export * from "./flow-type/flow-type.update.ts"
-export * from "./flow-type/flow-type.delete-request.ts"
 export * from "./flow-type/flow-type.request-delete.ts"
+export * from "./flow-type/flow-type.update.ts"
 
 // Event Types
+export * from "./event-type/event-type.create.ts"
+export * from "./event-type/event-type.delete-request.ts"
 export * from "./event-type/event-type.exists.ts"
 export * from "./event-type/event-type.fetch.ts"
 export * from "./event-type/event-type.list.ts"
-export * from "./event-type/event-type.create.ts"
-export * from "./event-type/event-type.update.ts"
-export * from "./event-type/event-type.delete-request.ts"
-export * from "./event-type/event-type.truncate-request.ts"
-export * from "./event-type/event-type.request-truncate.ts"
 export * from "./event-type/event-type.request-delete.ts"
+export * from "./event-type/event-type.request-truncate.ts"
+export * from "./event-type/event-type.truncate-request.ts"
+export * from "./event-type/event-type.update.ts"
 
 // Events
-export * from "./events/events.fetch.ts"
 export * from "./events/events.fetch-time-buckets-by-names.ts"
 export * from "./events/events.fetch-first-and-last-time-buckets.ts"
+export * from "./events/events.fetch.ts"
 
 //Container
 export * from "./container-registry/container-registry.create.ts"
+export * from "./container-registry/container-registry.delete.ts"
 export * from "./container-registry/container-registry.fetch.ts"
 export * from "./container-registry/container-registry.list.ts"
 export * from "./container-registry/container-registry.update.ts"
-export * from "./container-registry/container-registry.delete.ts"